+++ x/auth/genesis.go
@@ -3,13 +3,14 @@ package auth
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/exported"
+	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
 // InitGenesis - Init store state from genesis data
 //
 // CONTRACT: old coins from the FeeCollectionKeeper need to be transferred through
 // a genesis port script to the new fee collector account
-func InitGenesis(ctx sdk.Context, ak AccountKeeper, data GenesisState) {
+func InitGenesis(ctx sdk.Context, ak AccountKeeper, sk types.SupplyKeeper, data GenesisState) {
 	ak.SetParams(ctx, data.Params)
 	data.Accounts = SanitizeGenesisAccounts(data.Accounts)
 
@@ -17,6 +18,8 @@ func InitGenesis(ctx sdk.Context, ak AccountKeeper, data GenesisState) {
 		acc := ak.NewAccount(ctx, a)
 		ak.SetAccount(ctx, acc)
 	}
+
+	sk.GetModuleAccount(ctx, FeeCollectorName)
 }
 
 // ExportGenesis returns a GenesisState for a given context and keeper